Comments on CommonTools/RecoAlgos
