--- CommonTools/RecoAlgos/plugins/SealModule.cc
@@ -1,6 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "CommonTools/RecoAlgos/plugins/ME0MuonTrackCollProducer.h"
-
-DEFINE_FWK_MODULE(ME0MuonTrackCollProducer);
-
